# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.UpdateNotificationPreferencesRequest:
    properties:
      ai_health_tips_enabled:
        type: boolean
      app_updates_enabled:
        type: boolean
      appointment_reminders_enabled:
        type: boolean
      email_enabled:
        type: boolean
      medication_reminders_enabled:
        type: boolean
      push_enabled:
        type: boolean
      refill_reminders_enabled:
        type: boolean
      sms_enabled:
        type: boolean
      support_updates_enabled:
        type: boolean
      timezone:
        description: |-
          Timezone is an IANA name such as "Africa/Lagos". The reminder scheduler
          resolves it on every tick, so it is validated on write.
        maxLength: 64
        minLength: 1
        type: string
      whatsapp_enabled:
        type: boolean
    type: object
...
  request.UpdateUserRequest:
    properties:
      avatar_url:
        maxLength: 2048
        type: string
      blood_type:
        enum:
        - A+
        - A-
        - B+
        - B-
        - O+
        - O-
        - AB+
        - AB-
        type: string
      country_code:
        type: string
      dob:
        type: string
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      height:
        maximum: 300
        type: number
      last_name:
        maxLength: 100
        minLength: 1
        type: string
      sex:
        description: 'Sex: 1=male 2=female 3=other'
        maximum: 3
        minimum: 1
        type: integer
      temperature_unit:
        enum:
        - celsius
        - fahrenheit
        type: string
      timezone:
        description: |-
          Timezone is an IANA name such as "Africa/Lagos". Also settable on
          PATCH /users/me/notification-preferences; it is accepted here too because
          this is the endpoint a client reaches for when filling in a profile, and
          leaving it out meant new users silently kept the UTC default and received
          every reminder at the wrong local time.
        maxLength: 64
        minLength: 1
        type: string
      weight:
        maximum: 700
        type: number
      weight_unit:
        enum:
        - kg
        - lb
        type: string
    type: object
...
  response.APIResponse-response_EmptyData:
    properties:
      data:
        $ref: '#/definitions/response.EmptyData'
      errors: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
...
  response.APIResponse-response_GetUserResponse:
    properties:
      data:
        $ref: '#/definitions/response.GetUserResponse'
      errors: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
...
  response.APIResponse-response_NotificationPreferencesResponse:
    properties:
      data:
        $ref: '#/definitions/response.NotificationPreferencesResponse'
      errors: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
...
  response.EmptyData:
    type: object
...
  response.GetUserEmergencyContactDTO:
    properties:
      id:
        type: string
      is_primary:
        type: boolean
      name:
        type: string
      phone:
        type: string
      relationship:
        type: string
    type: object
  response.GetUserResponse:
    properties:
      avatar_url:
        type: string
      blood_type:
        type: string
      country:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      dob:
        type: string
      email:
        type: string
      emergency_contact:
        $ref: '#/definitions/response.GetUserEmergencyContactDTO'
      first_name:
        type: string
      height:
        type: number
      id:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      onboarding_completed:
        type: boolean
      phone:
        type: string
      sex:
        type: string
      status:
        type: string
      temperature_unit:
        type: string
      updated_at:
        type: string
      weight:
        type: number
      weight_unit:
        type: string
    type: object
...
  response.NotificationPreferencesResponse:
    properties:
      ai_health_tips_enabled:
        type: boolean
      app_updates_enabled:
        type: boolean
      appointment_reminders_enabled:
        type: boolean
      email_enabled:
        type: boolean
      medication_reminders_enabled:
        type: boolean
      push_enabled:
        type: boolean
      refill_reminders_enabled:
        type: boolean
      sms_enabled:
        type: boolean
      support_updates_enabled:
        type: boolean
      timezone:
        type: string
      whatsapp_enabled:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HTTP API for MediLog authentication, health records, medication tracking,
    AI conversations, drug verification, and reference data.
  title: MediLog API
  version: "1.0"
paths:
...
  /users/me:
    get:
      description: Returns the profile of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse-response_GetUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
      security:
      - BearerAuth: []
      summary: Get authenticated user's profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Partial update — omitted fields are left unchanged. Returns the
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse-response_GetUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
      security:
      - BearerAuth: []
      summary: Update the authenticated user's profile
      tags:
      - Users
  /users/me/notification-preferences:
    get:
      description: Returns the caller's delivery preferences and reminder timezone.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse-response_NotificationPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
      security:
      - BearerAuth: []
      summary: Get notification preferences
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Partial update — omitted toggles are left unchanged. The
      parameters:
      - description: Preferences to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateNotificationPreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse-response_NotificationPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.APIResponse-response_EmptyData'
      security:
      - BearerAuth: []
      summary: Update notification preferences
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"